New Base Set
Restarted the project with a new plan of action. "makefile" and
"host.h" was patched to fix the dependency issue and a missing
variable issue.

"host.c"
"int rcvflag" was added to the struct hoststate.

"makefile"
"*.h" files were added to the dependency for each component.

Next step is to add in a port number variable to the struct in
order to define the number of connections for each device. Hosts
will only have one port, labeled "port 0".